ref: move serializer code to serializer file, duh
